appFacturaDto/facturas/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary<3.0,>=2.9  # Cambiado a psycopg2-binary
gunicorn>=20.1,<21.0  # Servidor WSGI para producción
